// frontend/lib/types/enums.ts
export enum ModelType {
  order = "order",
  product = "product",
  category = "category",
}

export enum FormType {
  TEXT = "text",
  TEXTAREA = "textarea",
  NUMBER = "number",
  AutoComplete = "AutoComplete",
  Switch = "Switch",
  ImagesEditor = "ImagesEditor",
}

export enum OrderStatus {
  NEW = "new",
  READY = "ready",
  DONE = "done",
  CANCELED = "canceled",
}